This change adds support for event statements summary metrics
It does so by creating a new panel (with underlying module) that shows a top-like table with the following metrics
per query digest text:

* Number of queries.
* Query latency.
* Query lock time.
* Query CPU time.
* Rows examined.
* Rows affected.
* Rows sent.
* Query errors.
* Query warnings.
* Queries using bad indexes.
* Queries using no indexes.

The user may optionally toggle the display of the query digest text with that of a sample query with key `A`.

The source of data for this panel is performance_schema.events_statements_summary_by_digest. However, instead of
simply showing the metrics as they are available in the table, which refer statistics since MySQL started, it
shows them for dolphie's sampling period (given by `--refresh-interval`), thus showing a more current view of
the statements' performance, as opposed to a historical one.