This change adds support for event statements summary metrics (#91)
* This change adds support for event statements summary metrics

It does so by creating a new panel (with underlying module) that shows a top-like table with the following metrics
per query digest text:

* Number of queries.
* Query latency.
* Query lock time.
* Query CPU time.
* Rows examined.
* Rows affected.
* Rows sent.
* Query errors.
* Query warnings.
* Queries using bad indexes.
* Queries using no indexes.

The user may optionally toggle the display of the query digest text with that of a sample query with key `A`.

The source of data for this panel is performance_schema.events_statements_summary_by_digest. However, instead of
simply showing the metrics as they are available in the table, which refer statistics since MySQL started, it
shows them for dolphie's sampling period (given by `--refresh-interval`), thus showing a more current view of
the statements' performance, as opposed to a historical one.

* Fix AttributeError exception

This came from trying to access the module's data before it was fully initialized.

* Address PR comments:

* Format query digest & query sample.
* Fix legend of command to turn on/off panel.

* Fixing things I found wrong

* Misc

* Improve ordering of statements summary rows.

Order by last_seen desc in addition to sum_timer_wait to show most
recent queries first.

* Adding conversion of picseconds to human readable

* Statements Summary now does a cumulative diff from original value instead of per second diff. This makes the data more understandable

* Misc

* Use existing PerformanceSchemaMetrics module instead of creating a new one

* Cleanup

* Cleanup

* Add quantile 95/99

* Fix main branch conflict

---------

Co-authored-by: Charles Thompson <cthompson@digitalocean.com>